requirements.txt
fastapi
uvicorn
pydantic
python-multipart
transformers
torch
openvino
optimum-intel
nncf # for 4bit and 8bit quantization